requirements.txt
python-telegram-bot>=20.7
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
nest-asyncio>=1.6.0

